# code/orders-api-rest/api/openapi-rest.yml
openapi: 3.0.1
info:
  title: Retail API
  description: Retail service API
  version: 2.0.8
servers:
  - url: http://localhost:8080/retail

#TAGS
#=======================================================================================================================
...
paths:
...
  /v1/management/orders/{orderId}:
    get:
      tags:
        - OrdersManagement
      summary: Get order by id
      operationId: getOrderById
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
            description: Id of the order
            format: uuid
          example: '8efbee82-8a0c-407a-a4c0-16bbad40a23e'
        - $ref: '#/components/parameters/Language'
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagerOrder'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
  /v1/management/orders/{orderId}/status:
    patch:
      tags:
        - OrdersManagement
      summary: Update order statuses by order ID
      operationId: updateOrderStatus
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the order to update
          example: '8efbee82-8a0c-407a-a4c0-16bbad40a23e'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusRequest'
      responses:
        '200':
          description: Order status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderStatusInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
...
  /v1/management/orders:
    get:
      tags:
        - OrdersManagement
      summary: View all orders filtered and sorted
      description: Endpoint allows users to view all orders with optional filtering and sorting. It supports pagination, language localization, and requires bearer token authentication.
      operationId: getAllOrders
      security:
        - bearerAuth: [ ]
      parameters:
        - in: query
          name: ordersFilter
          required: false
          schema:
            $ref: '#/components/schemas/OrdersFilterParameters'
        - $ref: '#/components/parameters/Language'
        - $ref: '#/components/parameters/Pageable'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageManagerOrderPreview'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /v1/products/sales:
    get:
      tags:
        - Products
      summary: Get list of products on sale
      operationId: getProductsOnSale
      parameters:
        - in: query
          description: Filter for finding products on sale
          name: productFilter
          schema:
            $ref: '#/components/schemas/ProductsOnSaleFilter'
        - $ref: '#/components/parameters/Pageable'
          description: Sorting is available for discount/priceWithDiscount/name/percentageOfTotalOrders. The field 'percentageOfTotalOrders' represents the bestsellers. For example, sort=discount,desc or sort=name,asc.
        - $ref: '#/components/parameters/Language'
      responses:
        '200':
          description: Short information about all products which are currently on sale
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductsOnSaleResponse'
  /v1/products:
    get:
      tags:
        - Products
      summary: Get list of products
      description: The endpoint for getting all products from the database.
      operationId: getProducts
      parameters:
        - in: query
          description: filter for products
          name: productFilter
          schema:
            $ref: '#/components/schemas/ProductFilter'
        - $ref: '#/components/parameters/Pageable'
          description: For sorting by product id/status/createdAt/quantity/price you need to add "product." before field and name/description you need to send just name or description followed by ',' asc or desc. For example, sort=product.price,desc or sort=name,desc.
        - $ref: '#/components/parameters/Language'
      responses:
        '200':
          description: Short information about all products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageProducts'
  /v1/products/search:
    get:
      tags:
        - Products
      summary: Get page of products in search.
      description: The endpoint for getting page of short product info in search.
      operationId: searchProducts
      parameters:
        - in: query
          name: searchQuery
          required: true
          schema:
            minLength: 3
            pattern: '^(?!\s*$).+'
            type: string
            nullable: false
            example: phone
        - $ref: '#/components/parameters/Language'
        - $ref: '#/components/parameters/Pageable'
      responses:
        '200':
          description: Short information about all products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageProductSearchResult'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
  /v1/products/bestsellers:
    get:
      summary: Get the top bestselling products
      tags:
        - Products
      operationId: findBestsellers
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
          required: true
        - $ref: '#/components/parameters/Language'
      responses:
        '200':
          description: Successfully retrieved the bestsellers.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageProducts'
  /v1/products/bestseller:
    get:
      summary: Get a bestseller for a category in a period of time
      tags:
        - Products
      operationId: findMostSoldProductByTag
      x-spring-extras:
        responseType: ResponseEntity
      parameters:
        - $ref: '#/components/parameters/Language'
        - $ref: '#/components/parameters/Tag'
      responses:
        '200':
          description:  Returns the top-selling item in a given category. The response may optionally include a field indicating the percentage of sales over a specific time period.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPreview'
        '204':
          description: User deleted successfully. No content returned.
  /v1/products/{productId}:
    get:
      tags:
        - Products
      summary: Get product by ID
      description: Endpoint retrieves the details of an existing product identified by productId. It supports optional language localization.
      operationId: getProductDetailsById
      parameters:
        - $ref: '#/components/parameters/ProductId'
        - $ref: '#/components/parameters/Language'
      responses:
        '200':
          description: Successfully retrieved product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetailsResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
...
  /v1/management/users/{userId}/status:
    patch:
      tags:
        - UsersManagement
      operationId: changeUserStatus
      summary: Block/unblock user's account
      description: The endpoint change status of user's account
      parameters:
        - $ref: '#/components/parameters/UserId'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/AccountStatus'
      responses:
        '204':
          description: Status was updated
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
  /v1/management/users:
    get:
      tags:
        - UsersManagement
      summary: Get list of all users
      description: The endpoint for getting all accounts from the database.
      operationId: getAccounts
      parameters:
        - in: query
          description: filter for accounts
          name: accountFilter
          schema:
            $ref: '#/components/schemas/AccountFilter'
        - $ref: '#/components/parameters/Pageable'
      responses:
        '200':
          description: Short information about all accounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageAccounts'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
...
  parameters:
    Pageable:
      name: pageable
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/Pageable'
    UserId:
      in: path
      name: userId
      required: true
      description: Id of the user
      schema:
        type: integer
        format: int64
      example: '1'
...
    ProductId:
      in: path
      name: productId
      required: true
      description: ID of the product
      schema:
        type: string
        format: uuid
        example: '8efbee82-8a0c-407a-a4c0-16bbad40a23e'
    Language:
      in: query
      name: lang
      required: false
      schema:
        type: string
        default: uk
        nullable: false
        example: uk
      description: The language code for the product information (e.g., 'uk' for Ukrainian).
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      in: query
      name: size
      schema:
        type: integer
        minimum: 1
    Tag:
      in: query
      name: tag
      required: true
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  #RESPONSES
  #=======================================================================================================================
  responses:
...
    NotFoundResponse:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorObject'
          example:
            status: 404
            title: Not Found
            detail: The requested resource was not found
...
    BadRequestResponse:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorObject'
          example:
            status: 400
            title: Bad Request
            detail: Invalid Request. Please, check the data in the request (QueryString Parameters and/or Body).
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorObject'
          example:
            status: 401
            title: Unauthorized
            detail: Unauthorized access to the resource
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorObject'
          example:
            status: 403
            title: Forbidden
            detail: User does not have success to this resource

  #SCHEMAS
  #=======================================================================================================================
  schemas:
...
    Discount:
      type: integer
      example: 10
...
    ProductDetailsResponse:
      type: object
      properties:
        name:
          type: string
          example: "APPLE iPhone"
        description:
          type: string
          example: "Screen 6.1 Super Retina XDR, 2532x1170 / A16 Bionic chip"
        image:
          type: string
          example: "https://example.com/image.jpg"
        tags:
          type: array
          items:
            type: string
            example: "category:mobile, category:mobile"
        quantity:
          type: integer
          example: 10
        price:
          type: number
          example: 90.00
          description: Final price
        discount:
          $ref: '#/components/schemas/Discount'
          description: The amount of the discount applied to the original price. This field is only returned if a discount is applied.
        priceWithDiscount:
          type: number
          example: 100.00
          description: The original price before any discount is applied. This field is only returned if a discount is applied.
        percentageOfTotalOrders:
          description: The percentage of total orders for this product for a period of the last 30 days. The field can be absent.
          $ref: '#/components/schemas/Percentage'
...
    ManagerOrder:
      type: object
      description: Manager's view of Order
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the order
        isPaid:
          type: boolean
          description: Indicates whether the order has been paid
        orderStatus:
          $ref: '#/components/schemas/OrderStatus'
        createdAt:
          type: string
          format: date-time
          description: The date and time when the order was created
        editedAt:
          type: string
          format: date-time
          description: The date and time when the order was edited
        account:
          $ref: '#/components/schemas/AccountResponse'
        total:
          type: number
          format: decimal
          description: The total cost of the order
        totalWithDiscount:
          type: number
          format: decimal
          description: The total cost of the order, but at least one discount is applied. The field may not be returned.
        receiver:
          $ref: '#/components/schemas/OrderReceiver'
        postAddress:
          $ref: '#/components/schemas/PostAddress'
        orderItems:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    OrderItem:
      type: object
      properties:
        price:
          type: number
          format: decimal
          description: Price of the individual item
        priceWithDiscount:
          type: number
          format: decimal
          description: Price of the individual item if a discount is applied. The field may be absent
        discount:
          $ref: '#/components/schemas/Discount'
          description: Discount of the item when the user made an order. The field may be absent
        quantity:
          type: integer
          description: Quantity of the item ordered
        product:
          $ref: '#/components/schemas/ProductPreview'
    OrderReceiver:
      type: object
      properties:
        firstName:
          type: string
          description: Receiver's first name
        lastName:
          type: string
          description: Receiver's last name
        email:
          type: string
          description: Receiver's email address
    PostAddress:
      type: object
      properties:
        deliveryMethod:
          $ref: '#/components/schemas/DeliveryMethod'
        city:
          type: string
          description: City of the delivery address
        department:
          type: string
          description: Department or region of the delivery address
    OrderStatus:
      type: string
      enum:
        - IN_PROGRESS
        - SHIPPED
        - DELIVERED
        - COMPLETED
        - CANCELED
      description: The various statuses an order can have
      example: IN_PROGRESS, SHIPPED, DELIVERED, COMPLETED, CANCELED
    AccountResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Id of the account.
        email:
          type: string
          format: email
          example: example@mail.com
          description: Email of the account.
        firstName:
          type: string
          example: John
          description: First name of the account.
        lastName:
          type: string
          example: Doe
          description: Last name of the account.
    DeliveryMethod:
      type: string
      enum:
        - NOVA
        - UKRPOSHTA
      description: The delivery methods available for shipping the order
    OrdersFilterParameters:
      type: object
      properties:
        accountEmail:
          type: string
          example: example@mail.com
          description: Search by email starting with query.
        deliveryMethods:
          type: array
          items:
            $ref: '#/components/schemas/DeliveryMethod'
        statuses:
          type: array
          items:
            $ref: '#/components/schemas/OrderStatus'
        isPaid:
          type: boolean
          description: Indicates whether the order has been paid
        createdBefore:
          type: string
          format: date-time
          description: The date and time before the order was created
        createdAfter:
          type: string
          format: date-time
          description: The date and time after the order was created
        totalMore:
          type: number
          format: decimal
          description: The total cost of the order is more than
        totalLess:
          type: number
          format: decimal
          description: The total cost of the order is less than
...
    PageProductSearchResult:
      type: object
      allOf:
        - $ref: '#/components/schemas/AbstractPage'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/ProductSearchResult'
              description: List of product search results in page
    PageManagerOrderPreview:
      type: object
      allOf:
        - $ref: '#/components/schemas/AbstractPage'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/ManagerOrderPreview'
              description: List of orders in the page
    PageProducts:
      type: object
      allOf:
        - $ref: '#/components/schemas/AbstractPage'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/ProductPreview'
    ProductsOnSaleResponse:
      type: object
      properties:
        minimumPriceWithDiscount:
          $ref: '#/components/schemas/MinimumPriceWithDiscount'
          description: The lowest price after applying the discount to all products.
        maximumPriceWithDiscount:
          $ref: '#/components/schemas/MaximumPriceWithDiscount'
          description: The highest price after applying the discount to all products.
        minimumDiscount:
          $ref: '#/components/schemas/Discount'
          description: The lowest discount percentage applied to any product.
        maximumDiscount:
          $ref: '#/components/schemas/Discount'
          description: The highest discount percentage applied to any product.
        pageProducts:
          $ref: '#/components/schemas/PageProducts'
    ProductPreview:
      type: object
      properties:
        id:
          type: string
          example: '123e4567-e89b-12d3-a456-426614174000'
          description: Product Id.
        name:
          type: string
          example: APPLE iPhone
          description: Name of the product.
        description:
          type: string
          example: Device type, operation system, etc.
          description: Product description
        status:
          type: string
          default: AVAILABLE
          enum:
            - AVAILABLE
            - END_SOON
            - ENDED
          example: AVAILABLE
          description: Status of the product's quantity
        tags:
          type: array
          description: List of tags associated with the product
          items:
            type: string
            example: category:electronics
            description: Tag title
        image:
          type: string
          example: https://example.com/image.jpg
          description: Image URLs associated with the product.
        price:
          type: number
          example: 90.00
          description: Final price
        discount:
          type: integer
          example: 10
          description: The amount of the discount applied to the original price. This field is only returned if a discount is applied.
        priceWithDiscount:
          type: number
          example: 100.00
          description: The original price before any discount is applied. This field is only returned if a discount is applied.
        percentageOfTotalOrders:
          description: The percentage of total orders for this product for a period of the last 30 days. The field can be absent.
          $ref: '#/components/schemas/Percentage'
...
    ProductTags:
      type: array
      items:
        type: string
        description: categories of the product
        example: 'category:computer'
    UpdateOrderStatusRequest:
      type: object
      properties:
        orderStatus:
          $ref: '#/components/schemas/OrderStatus'
        isPaid:
          type: boolean
          description: Indicates if the order is paid
          example: true
      required:
        - orderStatus
    OrderStatusInfo:
      type: object
      properties:
        availableStatuses:
          type: array
          items:
            $ref: '#/components/schemas/OrderStatus'
        isPaid:
          type: boolean
          description: Indicates if the order is paid
          example: true
...
    ProductFilter:
      type: object
      properties:
        tags:
          $ref: '#/components/schemas/ProductTags'
    ProductsOnSaleFilter:
      type: object
      description: No fields required.
      properties:
        tags:
          $ref: '#/components/schemas/ProductTags'
        minimumDiscount:
          type: integer
          description: "The minimum discount percentage (from 0 to 100) to filter the products."
          example: 10
          minimum: 0
          maximum: 100
        maximumDiscount:
          type: integer
          description: "The maximum discount percentage (from 0 to 100) to filter the products."
          example: 50
          minimum: 0
          maximum: 100
        minimumPriceWithDiscount:
          $ref: '#/components/schemas/MinimumPriceWithDiscount'
        maximumPriceWithDiscount:
          $ref: '#/components/schemas/MaximumPriceWithDiscount'
...
    PageAccounts:
      type: object
      allOf:
        - $ref: '#/components/schemas/AbstractPage'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/AccountResponseManagement'
...
    AccountStatus:
      type: string
      enum:
        - ACTIVE
        - DEACTIVATED
      description: Status of the account
    AccountFilter:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/AccountStatus'
        role:
          $ref: '#/components/schemas/AccountManagementRoles'
...